@stephentoub 2019-08-22
> What do we expect the performance gains from this to be - for primary perf metrics like requests per second for something real looking?

Don't know yet. I'm hoping by putting this up as a draft folks will try it out :) I'm also going to try it on a few workloads. I won't be merging it until it proves valuable.