@benaadams 2019-08-22
Isn't very happy
```
Unhandled exception. System.InvalidOperationException: Operation is not valid due to the current state of the object.
   at System.Threading.Tasks.Sources.ManualResetValueTaskSourceCore`1.OnCompleted(Action`1 continuation, Object state, Int16 token, ValueTaskSourceOnCompletedFlags flags) in C:\GitHub\coreclr\src\System.Private.CoreLib\shared\System\Threading\Tasks\Sources\ManualResetValueTaskSourceCore.cs:line 163
   at System.Runtime.CompilerServices.AsyncValueTaskMethodBuilder`1.ValueTaskStateMachineBox.OnCompleted(Action`1 continuation, Object state, Int16 token, ValueTaskSourceOnCompletedFlags flags) in C:\GitHub\coreclr\src\System.Private.CoreLib\shared\System\Runtime\CompilerServices\AsyncValueTaskMethodBuilderT.cs:line 309
   at System.Runtime.CompilerServices.ValueTaskAwaiter`1.System.Runtime.CompilerServices.IStateMachineBoxAwareAwaiter.AwaitUnsafeOnCompleted(IAsyncStateMachineBox box) in C:\GitHub\coreclr\src\System.Private.CoreLib\shared\System\Runtime\CompilerServices\ValueTaskAwaiter.cs:line 188
   at System.Runtime.CompilerServices.AsyncTaskMethodBuilder`1.AwaitUnsafeOnCompleted[TAwaiter,TStateMachine](TAwaiter& awaiter, TStateMachine& stateMachine) in C:\GitHub\coreclr\src\System.Private.CoreLib\shared\System\Runtime\CompilerServices\AsyncTaskMethodBuilderT.cs:line 88
--- End of stack trace from previous location where exception was thrown ---
   at System.Threading.Tasks.Task.<>c.<ThrowAsync>b__139_1(Object state) in C:\GitHub\coreclr\src\System.Private.CoreLib\shared\System\Threading\Tasks\Task.cs:line 1910
   at System.Threading.QueueUserWorkItemCallbackDefaultContext.Execute() in C:\GitHub\coreclr\src\System.Private.CoreLib\shared\System\Threading\ThreadPool.cs:line 880
   at System.Threading.ThreadPoolWorkQueue.Dispatch() in C:\GitHub\coreclr\src\System.Private.CoreLib\shared\System\Threading\ThreadPool.cs:line 700
```