@benaadams 2019-08-22
> > Isn't very happy

> What is this running?

Just trying to debug into it as the stack trace clearly isn't very helpful :)