@halter73 2019-08-22
> e.g. if you were to await such a ValueTask multiple times, it would happen to work, even though we say "never do that". This change means that anyone who was relying on such undocumented behaviors will now be broken.

I say it's better to make this change sooner rather than later. I like that this change might teach people the hard way not to await the same ValueTask multiple times since it will no longer just Pipes and other IValueTaskSource-based ValueTasks that break.

Do you expect there will be any common workloads where pooling the boxes hurts perf?