@stephentoub 2019-08-22
> Do you expect there will be any common workloads where pooling the boxes hurts perf?

It's hard to say.

Using pooling is based on the assumption that the overheads associated with pooling (e.g. synchronization to access the pool, any cache effects that come from using the pool, etc.) are less than the overheads associated with allocation+GC.  That is not always the case, but it's also hard to predict.

There's also the fact that this is changing the underlying implementation, so it's not just pooling vs non-pooling, but ValueTask wrapping Task vs wrapping IValueTaskSource, different code paths being taken internally, different synchronization used to hook up a continuation, etc.  Some of that is just tuning, but some of it will be trade-offs as well.  In theory this approach could have less overhead even if pooling weren't used at all (just using a custom IValueTaskSource) because it doesn't need to accomodate things like multiple continuations, but we're also talking minutia at that point.

In the end, I'll want us to vet this on real workloads that demonstrate real value before moving ahead with it.