@mburbea 2019-08-22
@mgravell has a pretty interesting library that does something very similar to this 
https://github.com/mgravell/PooledAwait