@benaadams 2019-08-22
Revisiting https://github.com/dotnet/coreclr/issues/21612 again 😉 

Since `ValueTaskStateMachineBox` is a subtype would just calling it `StateMachineBox` work (as is done for many `Enumerator`s)?

So rather than 
`AsyncValueTaskMethodBuilder'1.ValueTaskStateMachineBox.OnCompleted`
It is 
`AsyncValueTaskMethodBuilder'1.StateMachineBox.OnCompleted`

However that's not too important 😄