@stephentoub 2019-08-22
> issue I'm having I think is in websockets

Yup, I wrote both that code and comment :)  Shame on me.  Will change it.