@stephentoub 2019-08-22
> not sure why its not really effected by this change

Probably some combination of 256 > than the Environment.ProcessCount*4 hardcoded cache size limit, plus the AsTask() in the WebSocket ReceiveAsync.