@stephentoub 2019-08-22
> not sure why its not really effected by this change

Oh, simpler explanation:
```C#
private async Task EnsureBufferContainsAsync
```
;-)