@benaadams 2019-08-22
I'm liking this very much. Even with its low(ish) pool limit the allocation growth is low

200 websocket connections ~30sec

![image](https://user-images.githubusercontent.com/1142958/63550427-8f00bf80-c52a-11e9-803e-d1abbffe875f.png)

1000 websocket connections ~30sec

![image](https://user-images.githubusercontent.com/1142958/63553491-772d3980-c532-11e9-9aa5-ed052f7129f5.png)

And the cost for renting and returning boxes seems fairly minimal