@stephentoub 2019-08-22
> I'm liking this very much

Thanks. Does it have a measurable impact on your throughout?