@benaadams 2019-08-22
The change in before and after allocations is huge 😄 

1000 websocket connections ~30sec

1.36M objects allocated (245MB)

![image](https://user-images.githubusercontent.com/1142958/63553974-08e97680-c534-11e9-9879-5d22ef2baad0.png)

209K objects allocated (20MB)

![image](https://user-images.githubusercontent.com/1142958/63554120-7d241a00-c534-11e9-9258-5ae0796d421b.png)